+++ pages/views/regulatory/listSDS.php
@@ -50,7 +50,7 @@ while($res = mysqli_fetch_array($q)){
     
 <!-- ADD NEW MODAL-->
 <div class="modal fade" id="createSDS" data-bs-backdrop="static" tabindex="-1" role="dialog" aria-labelledby="createSDS" aria-hidden="true">
-  <div class="modal-dialog modal-lg" role="document">
+  <div class="modal-dialog modal-xl" role="document">
     <div class="modal-content">
       <div class="modal-header">
         <h5 class="modal-title">Create SDS</h5>