+++ releasenotes.md
@@ -1,6 +1,7 @@
 Whats New in v10.5
 --------------------------
 - Introduse SDS creation (PREVIEW)
+- Added IFRA limit in composition if found in library
 - SDS in ingredient has been renamed to a Document
 - Added synonyms in document generation for ingredients
 - Auto create ingredient when importing a formula from Marketplace if not exists locally